+++ server/src/main/java/com/DH/server/service/implement/OrderServiceImpl.java
@@ -1,6 +1,8 @@
 package com.DH.server.service.implement;
 
+import com.DH.server.exception.EmailException;
 import com.DH.server.exception.OrderException;
+import com.DH.server.model.dto.EmailDTO;
 import com.DH.server.model.entity.Order;
 import com.DH.server.model.entity.Product;
 import com.DH.server.model.entity.UserEntity;
@@ -8,16 +10,22 @@ import com.DH.server.model.enums.ProductStatus;
 import com.DH.server.model.mapper.OrderMapper;
 import com.DH.server.persistance.OrderRepository;
 import com.DH.server.service.interfaces.AuthService;
+import com.DH.server.service.interfaces.EmailService;
 import com.DH.server.service.interfaces.OrderService;
 import com.DH.server.service.interfaces.ProductService;
+import jakarta.mail.MessagingException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -26,6 +34,7 @@ public class OrderServiceImpl implements OrderService {
   private final OrderMapper orderMapper;
   private final AuthService authService;
   private final ProductService productService;
+  private final EmailService emailService;
 
   @Override
   public Order create(Order entity, long productId) {
@@ -38,12 +47,15 @@ public class OrderServiceImpl implements OrderService {
       throw new OrderException("The product is unavailable, product id: "+ productId);
     }
     entity.setProduct(productTarget);
+    entity.setCreatedAt(LocalDateTime.now());
+    entity.setUser(authUser);
+
     List<Order> orders = this.orderRepository.getOrdersByProductIdAndDate(productId, LocalDate.now());
     this.verifyOverlapDates(orders, entity);
 
     long daysShip = ChronoUnit.DAYS.between(entity.getShipStart(), entity.getShipEnd());
     entity.setAmount(entity.getProduct().getPrice() * daysShip);
-    entity.setUser(authUser);
+
     return this.orderRepository.save(entity);
   }
 
@@ -96,6 +108,30 @@ public class OrderServiceImpl implements OrderService {
       if (!newEnd.isBefore(existingStart) && !newStart.isAfter(existingEnd)) {
         throw new OrderException("dates are overlapping");
       }
+    }
+  }
+  public void sendOrderConfirmation(UserEntity account, Order order){
+    long orderId = order.getId();
+    Map<String, String> variables = new HashMap<>();
+    variables.put("username", account.getName()+" "+account.getLastname());
+    variables.put("email", account.getEmail());
+    variables.put("OrderNumber", String.valueOf(orderId));
+
+    DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm:ss");
+    DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
+    variables.put("CreationDate", order.getCreatedAt().format(dateTimeFormatter));
+    variables.put("DeliveryDate", order.getShipEnd().format(dateFormatter));
+    variables.put("ShipAddress", order.getShipAddress());
+
+    Product product = order.getProduct();
+    variables.put("Product", product.getName());
+    variables.put("Amount", String.format("%.2f", order.getAmount()));
+
+    EmailDTO email = new EmailDTO(account.getEmail(), "Correo de confirmación de reservas", variables);
+    try {
+      this.emailService.sendMail(email, "emailOrder");
+    } catch (MessagingException e) {
+      throw new EmailException("Fail to send email to: "+account.getEmail());
     }
   }
 }
