+++ server/src/main/java/com/DH/server/service/interfaces/AuthService.java
@@ -2,7 +2,6 @@ package com.DH.server.service.interfaces;
 
 import com.DH.server.model.dto.request.LoginReq;
 import com.DH.server.model.dto.response.AuthRes;
-import com.DH.server.model.entity.Order;
 import com.DH.server.model.entity.UserEntity;
 
 public interface AuthService {
@@ -14,5 +13,4 @@ public interface AuthService {
 
   void resendEmailToken(String email);
 
-    void sendOrderConfirmation(UserEntity account, Order order);
 }
