+++ server/src/main/resources/templates/emailOrder.html
@@ -68,7 +68,6 @@
         valor:  [[${Amount}]]<br>
 
         Si tienes alguna pregunta o necesitas ayuda, no dudes en contactarnos.<br>
-        ¡Bienvenido a GameYard!<br>
         Saludos.<br>
         El equipo de GameYard</p>
 </main>
